Fixing additional compiler errors

+++ kaolin/csrc/metrics/sided_distance_cuda.cu
@@ -87,7 +87,7 @@ __global__ void sided_distance_forward_cuda_kernel(
             scalar_t z2 = buf[k * 3 + 2]- z1;
             scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-            if (k == 0 || __half2float(d) < __half2float(best)) {
+            if (k == 0 || d < best) {
               best = d;
               best_i = k + k2;
             }
@@ -99,7 +99,7 @@ __global__ void sided_distance_forward_cuda_kernel(
             scalar_t z2 = buf[k * 3 + 5] - z1;
             scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-            if (__half2float(d) < __half2float(best)){
+            if (d < best){
               best = d;
               best_i = k + k2 + 1;
             }
@@ -111,7 +111,7 @@ __global__ void sided_distance_forward_cuda_kernel(
             scalar_t z2 = buf[k * 3 + 8] - z1;
             scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-            if (__half2float(d) < __half2float(best)) {
+            if (d < best) {
               best = d;
               best_i = k + k2 + 2;
             }
@@ -123,7 +123,7 @@ __global__ void sided_distance_forward_cuda_kernel(
             scalar_t z2 = buf[k*3 + 11] - z1;
             scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-            if (__half2float(d) < __half2float(best)) {
+            if (d < best) {
               best = d;
               best_i = k + k2 + 3;
             }
@@ -137,7 +137,7 @@ __global__ void sided_distance_forward_cuda_kernel(
               scalar_t z2 = buf[k * 3 + 2] - z1;
               scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-              if (k == 0 || __half2float(d) < __half2float(best)) {
+              if (k == 0 || d < best) {
                 best = d;
                 best_i = k + k2;
               }
@@ -149,7 +149,7 @@ __global__ void sided_distance_forward_cuda_kernel(
               scalar_t z2 = buf[k * 3 + 5] - z1;
               scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-              if (__half2float(d) < __half2float(best)) {
+              if (d < best) {
                 best = d;
                 best_i = k + k2 + 1;
               }
@@ -161,7 +161,7 @@ __global__ void sided_distance_forward_cuda_kernel(
               scalar_t z2 = buf[k * 3 + 8] - z1;
               scalar_t d= x2 * x2 + y2 * y2 + z2 * z2;
 
-              if (__half2float(d) < __half2float(best)) {
+              if (d < best) {
                 best = d;
                 best_i = k + k2 + 2;
               }
@@ -173,7 +173,7 @@ __global__ void sided_distance_forward_cuda_kernel(
               scalar_t z2 = buf[k * 3 + 11] - z1;
               scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-              if (__half2float(d) < __half2float(best)) {
+              if (d < best) {
                 best = d;
                 best_i = k + k2 + 3;
               }
@@ -186,13 +186,13 @@ __global__ void sided_distance_forward_cuda_kernel(
           scalar_t z2 = buf[k * 3 + 2] - z1;
           scalar_t d = x2 * x2 + y2 * y2 + z2 * z2;
 
-          if (k == 0 || __half2float(d) < __half2float(best)) {
+          if (k == 0 || d < best) {
             best = d;
             best_i = k+k2;
           }
         }
 
-        if (k2 == 0 || __half2float(result[(i * n + j)]) > __half2float(best)) {
+        if (k2 == 0 || result[(i * n + j)] > best) {
           result[(i * n + j)] = best;
           result_i[(i * n + j)] = best_i;
         }
